Rename types/array.hpp to types/dynamic_array.hpp
Pure file rename; no content changes. The existing class is still named
Array; that rename happens in the next commit alongside the introduction
of the Array<T> dispatcher alias. types.hpp is updated to point at the
new path.
