+++ cpp/include/coconext/types.hpp
@@ -2,9 +2,9 @@
 #define COCONEXT_TYPES_HPP
 
 // NOLINTBEGIN(unused-includes)
-#include "./types/array.hpp"
 #include "./types/concepts.hpp"
 #include "./types/direction.hpp"
+#include "./types/dynamic_array.hpp"
 #include "./types/logic.hpp"
 #include "./types/range.hpp"
 // NOLINTEND(unused-includes)
